Move unpack to separate file to break cyclic dependency
Rollup gave warnings that the project contained cyclic
dependencies and serveral tests where failing mysteriously.
Fixed for now by moving unpack to it's own file.
Unsure if this is a good place to put it.
